split server_chunk in to separate file
