+++ shared/src/world/chunk/mod.rs
@@ -1,0 +1,5 @@
+mod chunk;
+mod server_chunk;
+
+pub use chunk::*;
+pub use server_chunk::*;
